Remove redundant pages for viewing/filtering data, move download page into existing research router
