+++ app/main.py
@@ -9,7 +9,7 @@ from fastapi.exceptions import RequestValidationError
 from starlette.middleware.sessions import SessionMiddleware
 from contextlib import asynccontextmanager
 
-from .routers import homepage, login, my_dreams, signup, global_stats, misc, research, research_data
+from .routers import homepage, login, my_dreams, signup, global_stats, misc, research
 from .utils import flash
 from .stat_calc import global_stat_calc_loop
 
@@ -37,7 +37,6 @@ app.include_router(login.router)
 app.include_router(signup.router)
 app.include_router(misc.router)
 app.include_router(research.router)
-app.include_router(research_data.router)
 
 @app.exception_handler(RequestValidationError)
 def validation_error_handler(request: Request, exc: RequestValidationError):
