+++ app/routers/__init__.py
@@ -1,2 +1,2 @@
-from . import homepage, login, my_dreams, signup, global_stats, misc, research_data
+from . import homepage, login, my_dreams, signup, global_stats, misc
