+++ app/templates/base.html
@@ -17,8 +17,8 @@
                 {% endif %}
                 {% if login_status == 2 %}
                     <a href="/research">Research filters</a> |
-                    <a href="/research/data">Research data</a> |
-                    <a href="/download-requested-data">Download CSV</a> |
+                    <a href="/research/data">View data</a> |
+                    <a href="/research/download">Download CSV</a> |
                 {% endif %}
                 <a href="/global-stats">Global stats</a> |
                 <a href="/about-us">About us</a>
