+++ app/templates/download-requested-data.html
@@ -21,8 +21,8 @@
   {% endif %}
 
   <p>
-    <a href="/request-data">Edit filters</a> |
-    <a href="/view-requested-data">View requested data</a>
+    <a href="/research">Edit filters</a> |
+    <a href="/research/data">View data online</a>
   </p>
 {% endblock %}
