--- app/templates/request-data.html
@@ -1,37 +1,0 @@
-{% extends "base.html" %}
-{% block body %}
-  <h1>Request data (research)</h1>
-  <p>Set filters for which <strong>public</strong> entries you want to access.</p>
-
-  <form method="post" action="/request-data">
-    <fieldset>
-      <legend>Date range (created_at)</legend>
-      <label for="start_date">Start date (YYYY-MM-DD):</label>
-      <input type="date" id="start_date" name="start_date" value="{{ filters.get('start_date') or '' }}">
-      <br>
-      <label for="end_date">End date (YYYY-MM-DD):</label>
-      <input type="date" id="end_date" name="end_date" value="{{ filters.get('end_date') or '' }}">
-    </fieldset>
-
-    <fieldset>
-      <legend>User location (exact match)</legend>
-      <label for="country">Country:</label>
-      <input type="text" id="country" name="country" value="{{ filters.get('country') or '' }}">
-      <br>
-      <label for="state">State:</label>
-      <input type="text" id="state" name="state" value="{{ filters.get('state') or '' }}">
-      <br>
-      <label for="city">City:</label>
-      <input type="text" id="city" name="city" value="{{ filters.get('city') or '' }}">
-    </fieldset>
-
-    <br>
-    <button type="submit">Save filters</button>
-  </form>
-
-  <p>
-    <a href="/view-requested-data">View requested data</a> |
-    <a href="/download-requested-data">Download requested data</a>
-  </p>
-{% endblock %}
-
