--- app/templates/view-requested-data.html
@@ -1,31 +1,0 @@
-{% extends "base.html" %}
-{% block body %}
-  <h1>View requested data (research)</h1>
-
-  {% if filters %}
-    <p><strong>Current filters:</strong> {{ filters }}</p>
-  {% endif %}
-
-  {% if entries|length == 0 %}
-    <p>No entries to display.</p>
-    <p><a href="/request-data">Set filters</a></p>
-  {% else %}
-    {% if limited %}
-      <p>Showing up to 200 newest matching entries.</p>
-    {% endif %}
-    <ul>
-      {% for e in entries %}
-        <li>
-          <strong>{{ e.title }}</strong> — {{ e.created_at.strftime('%Y-%m-%d') }}
-          <div style="white-space: pre-line;">{{ e.description }}</div>
-        </li>
-      {% endfor %}
-    </ul>
-  {% endif %}
-
-  <p>
-    <a href="/request-data">Edit filters</a> |
-    <a href="/download-requested-data">Download requested data</a>
-  </p>
-{% endblock %}
-
